+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sensors/gyro/Gyro.java
@@ -5,9 +5,9 @@ import org.firstinspires.ftc.teamcode.utils.Available;
 public interface Gyro extends Available {
     boolean isReady();
 
-    int getHeading();
+    double getHeading();
 
-    int getRaw();
+    double getRaw();
 
     void setOffset(int offset);
 }
